requirements.txt
PyQt6>=6.6.0
anthropic>=0.40.0
openai>=1.50.0
httpx>=0.27.0
html2text>=2024.2.26
Pillow>=10.0.0
openai